djangoLoan Service/requirements.txt
Django==3.2.13
environs==7.3.1
numpy-financial==1.0.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
